Backend/requirements.txt
aisuite==0.1.7
asyncpg==0.30.0
cerebras-cloud-sdk==1.15.0
cohere==5.12.0
fastapi>=0.109.1,<0.110.0
fastapi-analytics==1.2.3
google-search-results==2.4.2
langchain-core==0.3.28
groq
openai==1.58.1
psycopg2-binary==2.9.10
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv==1.0.0
sse-starlette==1.6.5
supabase==2.10.0
tavily-python==0.5.0
